+++ src/dac_jax/__init__.py
@@ -1,5 +1,8 @@
 __version__ = '1.1.0'
 
+__author__ = """David Braun"""
+__email__ = "braun@ccrma.stanford.edu"
+
 from dac_jax import nn
 from dac_jax import model
 from dac_jax import utils
