+++ src/dac_jax/audiotree/__init__.py
@@ -1,0 +1,1 @@
+from .core import AudioTree
